fix(download): break the helper-download module-init cycle (RUSH-3113)
